RSBackend/requirements.txt
APScheduler==3.9.0
Django==4.1
django_apscheduler==0.6.2
django-simpleui==2023.3.1
djangorestframework==3.14.0
freezegun==1.1.0
pytz==2023.3